adds refactoring for films and coments
