+++ src/const.js
@@ -1,0 +1,114 @@
+export const NAMES = [
+  'Tim Macoveev',
+  'John Doe',
+  'Tim Burry',
+  'Joe Johnson',
+  'Jimmy Klarkson',
+];
+
+export const COMMENTS = [
+  'Interesting setting and a good cast',
+  'Booooooooooring',
+  'Very very old. Meh',
+  'Almost two hours? Seriously?',
+];
+
+export const EMOJIS = [
+  'smile',
+  'sleeping',
+  'puke',
+  'angry',
+];
+
+export const TITLES = [
+  'Made For Each Other',
+  'Popeye Meets Sinbad',
+  'Sagebrush Trail',
+  'Santa Claus Conquers The Martians',
+  'The Dance Of Life',
+  'The Great Flamarion',
+  'The Man With The Golden Arm',
+];
+
+export const POSTERS = [
+  'made-for-each-other.png',
+  'popeye-meets-sinbad.png',
+  'sagebrush-trail.jpg',
+  'santa-claus-conquers-the-martians.jpg',
+  'the-dance-of-life.jpg',
+  'the-great-flamarion.jpg',
+  'the-man-with-the-golden-arm.jpg',
+];
+
+export const DIRECTORS = [
+  'Alfred Hitchcock',
+  'Stanley Kubrick',
+  'Akira Kurosawa',
+  'Steven Spielberg',
+  'Martin Scorsese',
+  'Quentin Tarantino',
+  'Ingmar Bergman',
+  'Federico Fellini',
+  'Francis Ford Coppola',
+  'David Lynch',
+];
+
+export const AUDIENCERATINGS = [
+  'G',
+  'PG',
+  'PG-13',
+  'R',
+  'NC-17',
+  '18+',
+];
+
+export const SENTENCES = [
+  'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
+  'Cras aliquet varius magna, non porta ligula feugiat eget.',
+  'Fusce tristique felis at fermentum pharetra.',
+  'Aliquam id orci ut lectus varius viverra.',
+  'Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante.',
+  'Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum.',
+  'Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui.',
+  'Sed sed nisi sed augue convallis suscipit in sed felis.',
+  'Aliquam erat volutpat.',
+  'Nunc fermentum tortor ac porta dapibus.',
+  'In rutrum ac purus sit amet tempus.',
+];
+
+export const GENRES = [
+  'Drama',
+  'Film-Noir',
+  'Mystery',
+  'Musical',
+  'Western',
+  'Drama',
+  'Comedy',
+  'Cartoon',
+];
+
+export const WRITERS = [
+  'Anne Wigton',
+  'Heinz Herald',
+  'Richard Weil',
+  'Jonathan Swift',
+  'Samuel Johnson',
+  'Johann Wolfgang von Goethe',
+  'Jane Austen',
+  'Alexandre Dumas',
+  'Victor Hugo',
+  'Charles Dickens',
+];
+
+export const ACTORS = [
+  'Robert De Niro',
+  'Jack Nicholson',
+  'Marlon Brando',
+  'Denzel Washington',
+  'Katharine Hepburn',
+  'Meryl Streep',
+  'Daniel Day-Lewis',
+  'Sidney Poitier',
+  'Clark Gable',
+  'Ingrid Bergman',
+];
